SimpleInterfaceMonitor filter events by bridge name
Add a filter in SimpleInterfaceMonitor. If provided, the Interface
events received will be filtered by port. For each Interface event
received, the bridge name will be retrieved and compared with the
list of bridge names given.

If no names are provided, the monitor won't filter any event
(current behavior).

This filter is meant to be used in Fullstack test only. This filtering
add an extra overload (two OVS DB queries per event) that should not
be needed in a production environment.

Closes-Bug: #1885547

Change-Id: Ie1fc8cf7d29c71eb358e593726b446787d8022c2