[EM releases] Move non-voting jobs to the experimental queue
During the CI meeting we agreed that non-voting jobs in the branches
which are in Extended Maintenance (EM) phase should be moved to from
the check to the experimental queue.

This patch is doing exactly that.

Conflicts:
    .zuul.yaml

Change-Id: Ie8a63eacf479ac6871af448a1741598584de8de8
(cherry picked from commit 1e2640fbb93c1b0afff7d5c079137e4e47587f5d)