[Functional] Add logging to the check test file function
In the functional tests module which is testing WSGI Server, there
is used function to assert that file created by the WSGI service is
created and has got correct size.
This function wasn't logging anything so in case of test failure
there was no way to check why it really failed.

This patch adds some log messages to the check so it will be clear
if file wasn't created at all or if size was not as expected.

Change-Id: Ib610b8513b4ea888a540873b26c3f205ac575b17
Related-Bug: #1886956