[ovn]: port forwarding -- documentation
This is a subset of the changes for implementing the floating IP
port forwarding feature in neutron, using OVN as the backend.

This changeset covers the documentation updates for the feature,
as well as a high-level description of how OVN implements it.

Partially-implements: ovn/port_forwarding
Partially-implements: blueprint portforwarding-description
Partial-Bug: #1877447

Change-Id: I2059a011f650dd7070a74dc6107aab2b15ca7104