[OVN] Extra DHCP options validation: Log invalid options
By raising an exception when an option is invalid we broke the Ironic +
OVN + Neutron DHCP agent combination that enabled deploying baremetal
machines.

This patch is changing the approach to just log the invalid options
instead of failing the request so that the OVN driver can still be used
with Ironic.

Change-Id: I5e98297acefb62f9a9c1200ccfaac0672eeeed2c
Closes-Bug: #1888649
Signed-off-by: Lucas Alvares Gomes <lucasagomes@gmail.com>