[ovn]: gracefully handle logical switch ports with tag set to None
AddLSwitchPortCommand takes kwargs as columns and tag is one of them.
It is possible that tag is provided with the value None and that is
not valid. The proposed change gracefully handles such cases so the
transaction is not aborted.

Change-Id: I0f3cb6f534072e166610b417059acbe89302473a
Closes-Bug: #1888736
(cherry picked from commit fe440bdd6be190588ec4310ef9630285c0acfec6)