Fix validation of IPv6 subnets with external RAs
In case when such subnet is added directly to the router, there was
validation and subnet which is expected to get RA messages from the
external router couldn't be added to the Neutron router.

But in case when port was first created manually and then plugged to the
router, there wasn't such validation. This patch fixes it by adding same
validation to adding router interface by port.

Change-Id: I054296c790b697198550acbeae29546758b422c2
Closes-Bug: #1889619