Use ebtables-legacy on Ubuntu Focal
On Ubuntu 20.04 there is by default ebtables-nft used. The problem with
that is that nftables don't supports syntax for source and destination
ipv4 address in arp tables right now.
So to workaround/fix this problem at least in our gate jobs we have to
switch to ebtables-legacy for now.
And that's exactly what this patch is doing.

Change-Id: I0aca119f1feca1d8f31a0104dde40d5d52c1b25c
Closes-Bug: #1889779