ovn: Always use UTC for Hash ring timestamps
When a node is created in ovn_hash_ring table, it used to use timestamp
depending on the system or mysql timezone. This caused troubles for
machines using western timezones when getting active nodes before the
first liveness check, becaus the time difference between UTC and local
time was always in hours and the limit is 30 seconds.

This patch configures OVN Hash Ring model to always use UTC when
creating a node so the times are always compared to UTC, regardless of
the used timezone. Note that the created_at and updated_at are not user
facing columns.

Conflicts:
	neutron/tests/unit/common/ovn/test_hash_ring_manager.py
	neutron/tests/unit/db/test_ovn_hash_ring_db.py

Change-Id: I9172254ce9b1880833128bdcaacacb93821933dd
Closes-bug: #1894117
Signed-off-by: Jakub Libosvar <libosvar@redhat.com>
(cherry picked from commit a84462f698f91b10e332e3403ecc9a68c3e6a884)