Update networking-midonet maintainers list
Ryu Ishimoto was removed from the list of networking-midonet maintainers
as he is not active in the community anymore - thank You Ryu for all
Your work in the Neutron community.

Sam Morrison was added to the list as he recently stepped in to maintain
networking-midonet in the Neutron stadium. Welcome Sam and thanks for
Your help with this project.

Change-Id: I639fa5f69f56c96bc69873e6bcac555fff441ce2