Remove extraneous garp emission
Previously, to work around a bug in keepalived, the
neutron-keepalived-state-change-monitor process was altered to emit
garps under certain circumstances.  That issue with keepalived issue
has been resolved, so remove the extraneous garp.

Two tests that tested the removed functionality have also been
removed.

This is a partial revert of https://review.opendev.org/393886/

Change-Id: I9fc5868067e890321ad880cc134fb034b5401f47
Co-Authored-By: Rodolfo Alonso Hernandez <ralonsoh@redhat.com>