[OVS] Remove sleep during polling cycle
OVS agent is a single threaded process with several eventlet threads
listening to RPC messages. During the polling cycle, the green thread
executing this code should not return or give the execution token to
other thread. Any other thread attending to RPC messages or OVSDB
events will be attended but should be processed at the end of the
cycle.

This patch removes a sleep introduced in the middle of a OF creation
and forces to have a sleep (even with time 0) at the end of the
polling cycle.

Related-Bug: #1905271

Change-Id: I14a9a212074a450d0e9ffd72dc05906d592542c5