Default dnsmasq --conf-file to /dev/null
Passing --conf-file= with no value has no effect on the dnsmasq
process. Intended effect here is for the default system dnsmasq.conf
file not to be read and included in configuring the process. For
that to happen some value has to be passed to --conf-file. Passing
/dev/null will invoke the desired outcome to skip the system
default conf file.

Closes-Bug: #1896945
Change-Id: I22570a44f84d14a792633747c04d7426ab231009
(cherry picked from commit 704576e54e041340ed9c2964b110815e074239a6)