ovn: support max_tunid option from NB_Global
This option is written by OVN and can be used to determine capacity of
the cluster that may depend on e.g. tunnel types enabled for
encapsulation.

The option is new in 20.09 and may be absent in older releases.  Hence
grace handling of missing option.

While this option is written by OVN regardless of encap types enabled,
it is useful only when VXLAN is enabled in the cluster. In this case,
due to limitations of the protocol header, the maximum number of
segments supported is drastically reduced from what's common when
Geneve is used.

The patch that adds VXLAN support for OVN is at:
I81c016ba9c91282d1bebb40a282077e14ce4bd6b

Change-Id: I8ddd52aa958f7dd07fee5d11913a498d6ad4f34f