Do not retrieve again the Floating IP object during creation
Floating IP OVO instance contains a reference to the DB object.
During a DB transaction (inside a session context), this reference
is always updated. There is no need to retrieve again the OVO
to read the updated version of the DB object.

For example, if a QoS policy is assigned to this Floating IP, the
DB object will update the "qos_policy_binding" reference.

Change-Id: Iec2552362f6c52842f12e20798324d4180d993e5
Related-Bug: #1880969