[ovn] Add 'net-mtu-writable' to supported extensions
This patch is adding the 'net-mtu-writable' extension to the
supported extensions list of the OVN plugin.

Without it, network creation with an MTU different from the
default (1500) will fail. Right now, devstack is working
because this extension gets loaded as a side effect of loading
'logger' but we should enable it by default.

Closes-Bug: #1901534

Depends-On: https://review.opendev.org/760446
Change-Id: I9ff47d8f678c95a6d6594f47ebca6175fdcc5453
Signed-off-by: Daniel Alvarez <dalvarez@redhat.com>