Remove neutron-tempest-{iptables_hybrid,linuxbridge} jobs from CI
This patch removes jobs:
neutron-tempest-linuxbridge
neutron-tempest-iptables_hybrid
from the Neutron check and gate queues.

Most of the test run there are not backend dependent and they are run in
the neutron-tempest-multinode-full-py3 job too.
Some tests which may be impacted by different backend/firewall driver
are added to the neutron-tempest-plugin-scenario jobs now.

Additionally this patch removes also
neutron-tempest-plugin-dvr-multinode-scenario from the docs summary as
this job is not run in the check queue for some time now (it's moved to
the experimental queue).

Depends-On: https://review.opendev.org/#/c/761410/
Change-Id: Ica669e05f719568dad74c4e8bd69ddceadccea7e