Disable cinder services on neutron grenade jobs
We don't need it to test cinder in our jobs.
This patch disables Cinder in grenade jobs. If that will be ok
we can do the same for scenario jobs also.

This patch also enables br-ex-tcpdump and br-int-flows
services in the grenade jobs as it was missing there and
may be useful during debugging some failures of those jobs
in the future.

Conflicts:
    zuul.d/grenade.yaml
    .zuul.yaml

Change-Id: If490da1f7e695564a65cabf95f25c72f7639ae19
(cherry picked from commit ca3f2813c26fb8a14c0853a4d6b4679088f7fbec)