Add support for vlan transparency in the OVN driver
This patch adds support for vlan_transparent in the ovn mechanism
driver. So ovn is now second mech_driver after linuxbridge which can be
used with vlan_transparent networks.
It just adds "vlan-passthru" option to the Logical Switch's "other
config".
This needs also changes in the core OVN which are available only in OVN
master branch for now. See [1] for details.

[1] https://patchwork.ozlabs.org/project/ovn/patch/20201110023449.194642-1-ihrachys@redhat.com/

Change-Id: I76b8ba959398dcffff112d26ae7d81ff428be992