Fix migration failed due TypeError of new_vlan_tag
There could be a situation, when there are ports
with port_tags None created in OS infra.

Under normal circumstances, it would take some error
to the log and go further.
But only because there is a "new_vlan_tags must be string"
in that log, it is throwing out typeError and dropping
the execution of code and the migration is ending by code 500.

And this is all done without a message from
the log, which is extremely bad.

This patch is fixing this behaviour, and flow
goes on.

Change-Id: Id4ba8d32e6b2301f25c12f83a0708b62aa5b9d6d
Closes-Bug: #1903989
(cherry picked from commit 7cf52bb5325ac73b7afa3212cb84d29831e3724a)