Disable not used services in the tempest and rally jobs
We don't need to run Swift, Cinder and etcd services in the
Neutron jobs so this patch disables those services on all
such jobs defined in the Neutron repository.
It will skip some tests which aren't really related to Neutron
and will safe some resources on test nodes.

This patch also enables br-ex-tcpdump and br-int-flows services
in the jobs where it was missing.
It may be useful during debugging some failures of those jobs
in the future.

Depends-On: https://review.opendev.org/762622

Change-Id: I0db516b184cf621e4539d11726fe1bd47b53ba05